Refactored the entire compilation part of qprogram⚛️
Also gates are now just matrices, with standard ones being just constant variables rather than return values of functions. (Performance should be slightly better now)
Compilation is refactored to deal with all this now, and is much cleaner. qprogram.compile() is now self contained, and doesn't depend on any other function.
At a high level the procedure to get a qprogram up and running is still the same (as in the example) (instantiate -> add gates -> compile -> run).
Next TODO: deal with the string representation of the qprogram.